Add breadcrumbs & minor ui adjustments
