+++ apps/admin/src/features/Curriculum/Lesson/Components/Editor/fields/index.ts
@@ -1,0 +1,4 @@
+export { ClozeGapAnswers } from "./ClozeGapAnswers";
+export { TriviaCorrectAnswer } from "./TriviaCorrectAnswer";
+export { AnalyzeCorrectOptions } from "./AnalyzeCorrectOptions";
+export { DistractorsEditor } from "./DistractorsEditor";
